S[ring profile to condition the Logging
Creating Spring profiles dev and prod and condition the logging to it. Only log to cloudwatch from the prod profile and only log to STO from the dev profile.
Passing the AWS key and secret to the image via arguments.
Getting the aws key and secret from env variables to build images locally.
Fixing github actions.